requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.60.0
scipy>=1.10.0
matplotlib>=3.5.0
